requirements-testing.txt
pytest
pytest-celery
tox
